requirements.txt
asn1crypto==1.5.1
attrs==22.2.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.1
cryptography==38.0.4
deepdiff==6.2.3
filelock==3.9.0
Flask==2.2.2
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-bigquery==3.4.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.1
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kafka-python==2.0.2
MarkupSafe==2.1.1
ordered-set==4.1.0
orjson==3.8.4
oscrypto==1.3.0
packaging==21.3
proto-plus==1.22.1
protobuf==4.21.12
psycopg==3.1.7
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.16.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
snowflake-connector-python==2.9.0
swagger-ui-bundle==0.0.9
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
